requirements.txt
astor==0.8.1
async-generator==1.10
av==8.0.2
certifi==2020.12.5
cffi==1.14.0
cycler==0.10.0
easydict==1.9
entrypoints==0.3
ffmpeg==1.4
google-pasta==0.2.0
imageio==2.8.0
ipython==7.8.0
jedi==0.17.0
jupyter==1.0.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
MulticoreTSNE==0.1
numpy==1.16.0
opencv-python==4.4.0.42
pandas==1.0.1
path==15.0.0
Pillow==7.0.0
prompt-toolkit==2.0.10
protobuf==3.14.0
pytz==2019.3
PyYAML==5.4.1
pyzmq==20.0.0
QtPy==1.9.0
scikit-learn==0.21.3
seaborn==0.10.0
style==1.1.0
tensorboard==1.14.0
tensorboardX==2.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.4
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
traitlets==4.3.3
update==0.0.1
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
